fix load of active tasks

+++ app/src/main/java/com/example/wizardlydo/viewmodel/TaskViewModel.kt
@@ -63,7 +63,7 @@ class TaskViewModel(
                 }
 
                 val tasks = runCatching {
-                    taskRepository.getAllTasks(userId.toInt())
+                    taskRepository.getAllTasks(userId)
                 }.fold(
                     onSuccess = { it },
                     onFailure = {
